+++ js/profile/components/my-hardware.js
@@ -8,7 +8,7 @@ import {
   parseSteamSystemInfo, inferGpuVendor,
   getMyHwSourceMeta, setMyHwSourceMeta, getMyHwFieldOrigins,
   setMyHwFieldOrigins, setMyHwFieldOrigin,
-} from '../utils.js?v=a77e8ddf';
+} from '../utils.js?v=38d77229';
 
 /**
  * Initialise the My Hardware pane. Call once after DOM is ready.